Refactor: 이슈 상태, 우선순위, 타입 enums 패키지로 이동
