+++ src/main/java/com/uranus/taskmanager/api/issue/domain/enums/IssueType.java
@@ -1,4 +1,4 @@
-package com.uranus.taskmanager.api.issue.domain;
+package com.uranus.taskmanager.api.issue.domain.enums;
 
 public enum IssueType {
 	EPIC,
