+++ src/main/java/com/uranus/taskmanager/api/issue/presentation/dto/request/UpdateStatusRequest.java
@@ -1,6 +1,6 @@
 package com.uranus.taskmanager.api.issue.presentation.dto.request;
 
-import com.uranus.taskmanager.api.issue.domain.IssueStatus;
+import com.uranus.taskmanager.api.issue.domain.enums.IssueStatus;
 
 import jakarta.validation.constraints.NotNull;
