+++ src/main/java/com/uranus/taskmanager/api/issue/presentation/dto/response/CreateIssueResponse.java
@@ -5,9 +5,9 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 
 import com.uranus.taskmanager.api.issue.domain.Issue;
-import com.uranus.taskmanager.api.issue.domain.IssuePriority;
-import com.uranus.taskmanager.api.issue.domain.IssueStatus;
-import com.uranus.taskmanager.api.issue.domain.IssueType;
+import com.uranus.taskmanager.api.issue.domain.enums.IssuePriority;
+import com.uranus.taskmanager.api.issue.domain.enums.IssueStatus;
+import com.uranus.taskmanager.api.issue.domain.enums.IssueType;
 
 public record CreateIssueResponse(
 	Long issueId,
